+++ src/cjs/Bounds.js
@@ -17,6 +17,7 @@
  * @modified 2025-10-17 Added the methods `Bounds.getSouthWestPoint`, `getNorthWestPoint`, `getNorthEastPoint` and `getSouthEastPoint`.
  * @modified 2025-10-18 Added method `Bounds.containsVert(XYCoords)`.
  * @modified 2025-10-28 Added the `Bounds.getWidth()` and `Bounds.getHeight()`.
+ * @modified 2025-11-12 Added the `Bounds.getMoved()` method.
  * @version  1.11.0
  **/
 Object.defineProperty(exports, "__esModule", { value: true });
@@ -62,6 +63,17 @@ var Bounds = /** @class */ (function () {
     };
     Bounds.prototype.getHeight = function () {
         return this.max.y - this.min.y;
+    };
+    Bounds.prototype.getScaled = function (scaleFactor) {
+        var center = this.getCenter();
+        var newMin = new Vertex_1.Vertex(this.min).scale(scaleFactor, center);
+        var newMax = new Vertex_1.Vertex(this.max).scale(scaleFactor, center);
+        return new Bounds(newMin, newMax);
+    };
+    Bounds.prototype.getMoved = function (moveAmount) {
+        var newMin = new Vertex_1.Vertex(this.min).move(moveAmount);
+        var newMax = new Vertex_1.Vertex(this.max).move(moveAmount);
+        return new Bounds(newMin, newMax);
     };
     /**
      * Get the center point of the north bound.