+++ changelog.md
@@ -1,5 +1,14 @@
 # Changelog
 
+- 2025-11-14
+  - Fixing a bug in the `drawutilssvg` class in the CSS `mix-blend-mode` property handling (caused a runtime error).
+  - Adding utility class `utils/datastructures/GeometryMesh`. Extending demo `33-depth-mesh` with the new class.
+  - Moved experimental class `Vert3` from demo 33 one level up.
+- 2025-11-12
+  - Added the `PerlinNoise` class and refactored, copied from a different project `floatsim`.
+  - Added the `Bounds.getMoved()` method.
+  - Added Demo `66-simple-perlin-noise`.
+  - Extended demo `64-multiple-ring-animation`: added perlin noise element.
 - 2025-10-29
   - Ported the Lissajous implementation from demo-13 to a Typescript class `utils/LissajousFigure`.
 - 2025-10-28